feat: add gamification setup
