+++ services/apiService.ts
@@ -1,3 +1,5 @@
+import { hexToDecimal } from "../utils/feltService";
+
 type LeaderboardTopperParams = {
   addr: string;
 };
